init main
