+++ Grid.py
@@ -1,11 +1,6 @@
 import numpy as np
 from enum import Enum
 
-class Direction(Enum):
-    UP = 0
-    RIGHT = 1
-    DOWN = 2
-    LEFT = 3 # cw
 
 class Grid:
 
@@ -104,54 +99,3 @@ class Grid:
             print("taux remplissage requis : " + str(self.tr))
             
         
-        
-
-
-grid_instance = Grid(10, 10,0.20)
-grid_instance.initGrid()
-grid_instance.setValue(1,1,2)
-grid_instance.setValue(2,1,2)
-grid_instance.setValue(2,2,2)
-grid_instance.setValue(2,3,2)
-grid_instance.setValue(3,3,2)
-grid_instance.setValue(3,4,2)
-grid_instance.setValue(3,5,2)
-grid_instance.setValue(4,5,2)
-grid_instance.setValue(5,5,2)
-grid_instance.setValue(5,6,2)
-grid_instance.setValue(5,7,2)
-grid_instance.setValue(4,7,2)
-grid_instance.setValue(3,7,2)
-grid_instance.setValue(2,7,2)
-grid_instance.setValue(2,6,2)
-grid_instance.setValue(1,6,2)
-print(grid_instance.grid)
-grid_instance.setEndPath(True)
-grid_instance.update(1,4,5,2)
-print(grid_instance.grid)
-print("Victoire? "+str(grid_instance.endGame))
-
-grid_instance2 = Grid(10, 10,0.30)
-grid_instance2.initGrid()
-grid_instance2.setValue(1,1,2)
-grid_instance2.setValue(2,1,2)
-grid_instance2.setValue(2,2,2)
-grid_instance2.setValue(2,3,2)
-grid_instance2.setValue(3,3,2)
-grid_instance2.setValue(3,4,2)
-grid_instance2.setValue(3,5,2)
-grid_instance2.setValue(4,5,2)
-grid_instance2.setValue(5,5,2)
-grid_instance2.setValue(5,6,2)
-grid_instance2.setValue(5,7,2)
-grid_instance2.setValue(4,7,2)
-grid_instance2.setValue(3,7,2)
-grid_instance2.setValue(2,7,2)
-grid_instance2.setValue(2,6,2)
-grid_instance2.setValue(1,6,2)
-print(grid_instance2.grid)
-grid_instance2.setEndPath(True)
-grid_instance2.update(4,4,5,2)
-print(grid_instance2.grid)
-print("Victoire? "+str(grid_instance2.endGame))
-
